Introduce per-page iteration
Similar style as Python is doing. As the consequence of this --
and the existing behavior probably confuses users -- page-streaming
methods always return a stream, and the callback is simply ignored.

The callback argument is kept for the compatibility.

Also noticed that test/api_callable.js is using assertion while
other tests are BDD style, fixed the style as well.