Refactor createApiCall
Currently createApiCall assumes to receive a function and
a CallSettings instance, and then returns a function.

On the reality, however, the API function can't be a normal
function but a Promise to a function (due to the asynchronous-ness
of auth library) and therefore, it's hard to handle the return
value of the code upon usage.

With this refactoring, the generated code can simply invokes
createApiCall with a Promise stub, and it simplifies the return
values.

This is a preparation to #20, which requires a different type of
return value to non-page-streaming functions, and #22 which will
also return a value.

This also reduces the size of the generated code because the
pattern used in the API call can be simplified. See
https://github.com/jmuk/toolkit/commit/ec23b2508a9e32350c803fcdf003acb5a1847214
to see how it's simplified.