Refactor gRPC and Auth situation.
As we discussed in https://github.com/GoogleCloudPlatform/gcloud-node/issues/1463',
many of auth parameters will be called in a initialization
of the module itself and then used later.

For this purpose, GAX also serves an object which keeps
the auth data and handle its feature under the same context.

This also offers a new dependency of google-auto-auth library
instead of google-auth-library, and supports nice optional parameters
like keyFile.