Remove CallOptions as a class.
I've realized that using just an Object is much more friendly to
JS programmers instead of having a class. This removes CallOptions
as a class, but the same interface is used as an Object.

Also removed FIRST_PAGE constant. Rather than specifying an object
like this, it would be more understandable to JS programmers to
have a boolean flag.

Updates #35