make metadata faster
Instead of creating fresh metadata every time,
we keep a copy-on-write instance.
In the hopefully-common case that the users don't need
to insert any metadata, we don't need to create new ones.

This commit also fixes a bug related to x-goog-api-client header.
Previously we checked the key string against "x-goog-api-client"
and refuse to insert the value if the key matches.
This is incorrect because header key is case-insensitive;
by making any character upper case, the key-value pair can be overwritten.
This commit correctly compares ignoring case.