test: make tests work in Node 15 (#930)
Node 15 introduced `AbortController` (so the test that mocks it now needs to be aware of which `AbortController` to mock - the Node 15 one or the one from `abort-controller` module).

Also, Node 15 apparently broke something in stream handling (the order of events coming has changed). Fixed the test to account for that possible change.